# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Press Kits Service",
    "description": "API for generating, versioning, and publicly sharing company press kits",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://press-kits.mcpfactory.org"
    }
  ],
  "components": {
    "schemas": {
...
      "MediaKitListResponse": {
        "type": "object",
        "properties": {
          "mediaKits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaKitSummary"
            }
          }
        },
        "required": [
          "mediaKits"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Media kit not found"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "UpdateMdxRequest": {
        "type": "object",
        "properties": {
          "mdxContent": {
            "type": "string",
            "example": "# Updated Press Kit\n\nNew content goes here..."
          }
        },
        "required": [
          "mdxContent"
        ]
      },
...
      "CreateMediaKitRequest": {
        "type": "object",
        "properties": {
          "mediaKitId": {
            "type": "string",
            "format": "uuid",
            "description": "Target a specific media kit. If omitted, the latest active kit in the scope (org + brand + campaign) is used. If no kit exists, a new one is created from scratch.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "instruction": {
            "type": "string",
            "description": "User instruction for the generation workflow. If omitted, generation proceeds using brand data only.",
            "example": "Create a professional press kit highlighting our Q1 2026 product launches and sustainability initiatives"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "paths": {
...
    "/media-kits": {
      "get": {
        "summary": "List media kits",
        "description": "Returns active media kits (validated, drafted, generating). At least one filter is required. Results are sorted by status priority (validated first) then by most recently updated.",
        "tags": [
          "Media Kits"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Filter by org UUID. At least one of org_id, campaign_id, or brand_id is required.",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": false,
            "name": "org_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Case-insensitive title search (partial match)",
              "example": "Q1 2026"
            },
            "required": false,
            "name": "title",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by campaign UUID",
              "example": "c7d8e9f0-a1b2-3456-cdef-789012345678"
            },
            "required": false,
            "name": "campaign_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by brand UUID (finds kits that include this brand)",
              "example": "a6b5fdad-b31d-4fa2-b34b-1cec4cb21ce5"
            },
            "required": false,
            "name": "brand_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Media kits list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or edit a media kit (versioning via fork)",
        "description": "Triggers a generation workflow for a media kit. The behavior depends on the current state:\n\n**No existing kit in scope** → Creates a new kit with `status: generating` and `parentMediaKitId: null`.\n\n**Existing kit is `validated` or `drafted`** → **Forks**: creates a NEW kit (new UUID) with `status: generating`, copying the content (title, icon, MDX) from the original. The new kit's `parentMediaKitId` points to the original, preserving full version history. The original kit is NOT modified or archived at this stage.\n\n**Existing kit is `generating`** → Updates the existing generating kit in place (adds the instruction, refreshes context headers). No new kit is created.\n\nScope is determined by `x-org-id` + `x-brand-id` + `x-campaign-id` headers. Pass `mediaKitId` in the body to target a specific kit instead of auto-resolving by scope.\n\nTo validate the generated result and archive the previous version, call `POST /media-kits/{id}/validate`. To discard the generating kit and restore the parent, call `POST /media-kits/{id}/cancel`.",
        "tags": [
          "Media Kits"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand UUID(s) — comma-separated list of brand UUIDs for multi-brand campaigns",
              "example": "a6b5fdad-b31d-4fa2-b34b-1cec4cb21ce5,b7c6gebe-c42e-5gb3-c45c-2ded5dc32df6"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign UUID — scopes kit to a specific campaign"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug (e.g. 'press-kit-v2')"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug override for generation"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Stable feature dynasty slug (version-independent). Forwarded to downstream services."
            },
            "required": false,
            "name": "x-feature-dynasty-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Stable workflow dynasty slug (version-independent). Forwarded to downstream services."
            },
            "required": false,
            "name": "x-workflow-dynasty-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaKitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Media kit created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Media kit not found (when mediaKitId specified)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/media-kits/{id}": {
      "get": {
        "summary": "Get media kit by ID",
        "description": "Returns full media kit including MDX content. Kits stuck in 'generating' for over 30 minutes are automatically transitioned to 'failed'.",
        "tags": [
          "Media Kits"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Media kit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/media-kits/{id}/mdx": {
      "patch": {
        "summary": "Update MDX content",
        "description": "Directly update the MDX page content of a media kit. Typically used for manual edits after generation.",
        "tags": [
          "Media Kits"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMdxRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/media-kits/{id}/status": {
      "patch": {
        "summary": "Update media kit status",
        "description": "Manually update a kit's status. Use 'denied' with a denialReason to reject a kit for rework.",
        "tags": [
          "Media Kits"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/media-kits/{id}/validate": {
      "post": {
        "summary": "Validate media kit",
        "description": "Approves the kit and sets it to 'validated'. Automatically archives any previously validated kit in the same scope (org + campaign). Sends a press_kit_ready email notification.",
        "tags": [
          "Media Kits"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}